Updated enemy movement, and changed some variable names to be clearer. Upgraded to unity 5.1

+++ Assets/Scripts/GameManager.cs
@@ -33,7 +33,7 @@ public class GameManager : MonoBehaviour {
 		UpdateMoneyImage();
 	}
 
-	public void TakeDamage(int amount) {
+	public void TakeDamage(float amount) {
 		curHealth -= amount;
 		UpdateHealthImage();
 	}
